Moved generic and light models to a dedicated files
* mplemented common methods for set/get actions
* moved light messages from generic directory
* removed generic and light models from models.py
