--- bluetooth_mesh/messages/generic/light/__init__.py
@@ -1,20 +1,0 @@
-# pylint: disable=W0611
-
-from .ctl import (
-    LightCTLMessage,
-    LightCTLOpcode,
-    LightCTLSetupMessage,
-    LightCTLSetupOpcode,
-)
-from .lightness import (
-    LightLightnessMessage,
-    LightLightnessOpcode,
-    LightLightnessSetupMessage,
-    LightLightnessSetupOpcode,
-)
-from .hsl import (
-    LightHSLMessage,
-    LightHSLOpcode,
-    LightHSLSetupMessage,
-    LightHSLSetupOpcode,
-)
