+++ bluetooth_mesh/messages/light/__init__.py
@@ -1,0 +1,14 @@
+from construct import Select
+
+from .ctl import LightCTLMessage, LightCTLSetupMessage
+from .lightness import LightLightnessMessage,LightLightnessSetupMessage
+from .hsl import LightHSLMessage, LightHSLSetupMessage
+
+LightMessage = Select(
+    LightCTLMessage,
+    LightCTLSetupMessage,
+    LightLightnessMessage,
+    LightLightnessSetupMessage,
+    LightHSLMessage,
+    LightHSLSetupMessage,
+)
