+++ bluetooth_mesh/models/light/__init__.py
@@ -1,0 +1,5 @@
+# pylint: disable=W0614
+
+from .lightness import *
+from .ctl import *
+from .hsl import *
