+++ bluetooth_mesh/models/models.py
@@ -39,23 +39,6 @@ from bluetooth_mesh.messages.config import (
     SecureNetworkBeacon,
     StatusCode,
 )
-from bluetooth_mesh.messages.generic.level import GenericLevelOpcode
-from bluetooth_mesh.messages.generic.light.ctl import LightCTLOpcode
-from bluetooth_mesh.messages.generic.light.lightness import (
-    LightLightnessOpcode,
-    LightLightnessSetupOpcode,
-)
-from bluetooth_mesh.messages.generic.light.hsl import (
-    LightHSLOpcode,
-    LightHSLSetupOpcode,
-)
-from bluetooth_mesh.messages.generic.onoff import GenericOnOffOpcode
-from bluetooth_mesh.messages.generic.dtt import GenericDTTOpcode
-from bluetooth_mesh.messages.generic.ponoff import (
-    GenericPowerOnOffOpcode,
-    GenericPowerOnOffSetupOpcode,
-    GenericOnPowerUp,
-)
 from bluetooth_mesh.messages.health import HealthOpcode
 from bluetooth_mesh.messages.properties import PropertyID
 from bluetooth_mesh.messages.scene import SceneOpcode
@@ -89,24 +72,8 @@ __all__ = [
     "DebugClient",
     "NetworkDiagnosticClient",
     "NetworkDiagnosticSetupClient",
-    "GenericOnOffServer",
-    "GenericOnOffClient",
-    "GenericDTTServer",
-    "GenericDTTClient",
-    "GenericPowerOnOffServer",
-    "GenericPowerOnOffSetupServer",
-    "GenericPowerOnOffClient",
     "SceneClient",
-    "GenericLevelClient",
     "SensorClient",
-    "LightLightnessClient",
-    "LightLightnessServer",
-    "LightCTLClient",
-    "LightHSLClient",
-    "LightHSLServer",
-    "LightHSLSetupServer",
-    "LightHSLHueServer",
-    "LightHSLSaturationServer",
     "GatewayConfigServer",
     "GatewayConfigClient",
     "LightExtendedControllerSetupClient",
@@ -1309,426 +1276,6 @@ class NetworkDiagnosticSetupClient(Model):
     PUBLISH = True
 
 
-class GenericOnOffServer(Model):
-    MODEL_ID = (None, 0x1000)
-    OPCODES = {
-        GenericOnOffOpcode.GENERIC_ONOFF_GET,
-        GenericOnOffOpcode.GENERIC_ONOFF_SET,
-        GenericOnOffOpcode.GENERIC_ONOFF_SET_UNACKNOWLEDGED,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-
-class GenericOnOffClient(Model):
-    MODEL_ID = (None, 0x1001)
-    OPCODES = {
-        GenericOnOffOpcode.GENERIC_ONOFF_STATUS,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-    async def set_onoff(
-        self,
-        destination: int,
-        app_index: int,
-        onoff: int,
-        delay: float = 0.5,
-        send_interval: float = 0.07,
-        timeout: Optional[float] = None,
-    ) -> int:
-        current_delay = delay
-        tid = self.tid()
-
-        status_opcode = GenericOnOffOpcode.GENERIC_ONOFF_STATUS
-
-        status = self.expect_app(
-            destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        async def request():
-            nonlocal current_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericOnOffOpcode.GENERIC_ONOFF_SET,
-                params=dict(
-                    onoff=onoff,
-                    tid=tid,
-                    transition_time=0,
-                    delay=current_delay,
-                ),
-            )
-            current_delay = max(0.0, current_delay - send_interval)
-
-            return await ret
-
-        status = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 0.5,
-        )
-
-        return status[status_opcode.name.lower()]["present_onoff"]
-
-    async def set_onoff_unack(
-        self,
-        destination: int,
-        app_index: int,
-        onoff: int,
-        delay: float = 0.5,
-        send_interval: float = 0.07,
-        transition_time: float = 0,
-        retransmissions: int = 6,
-    ):
-        current_delay = delay
-        tid = self.tid()
-
-        async def request():
-            nonlocal current_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericOnOffOpcode.GENERIC_ONOFF_SET_UNACKNOWLEDGED,
-                params=dict(
-                    onoff=onoff,
-                    tid=tid,
-                    transition_time=transition_time,
-                    delay=current_delay,
-                ),
-            )
-            current_delay = max(0.0, current_delay - send_interval)
-
-            return await ret
-
-        await self.repeat(
-            request,
-            retransmissions=retransmissions,
-            send_interval=send_interval,
-        )
-
-    async def get_onoff_status(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=GenericOnOffOpcode.GENERIC_ONOFF_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = GenericOnOffOpcode.GENERIC_ONOFF_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-
-class GenericDTTServer(Model):
-    MODEL_ID = (None, 0x1004)
-    OPCODES = {
-        GenericDTTOpcode.GENERIC_DTT_SET,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-
-class GenericDTTClient(Model):
-    MODEL_ID = (None, 0x1005)
-    OPCODES = {
-        GenericDTTOpcode.GENERIC_DTT_STATUS,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-    async def set_dtt(
-        self,
-        destination: int,
-        app_index: int,
-        transition_time: float,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> float:
-        status_opcode = GenericDTTOpcode.GENERIC_DTT_STATUS
-        status = self.expect_app(
-            destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        async def request():
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericDTTOpcode.GENERIC_DTT_SET,
-                params=dict(
-                    transition_time=transition_time,
-                ),
-            )
-
-            return await ret
-
-        status = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 0.5,
-        )
-
-        return status[status_opcode.name.lower()]["transition_time"]
-
-
-    async def set_dtt_unack(
-        self,
-        destination: int,
-        app_index: int,
-        transition_time: float,
-        *,
-        send_interval: float = 0.1,
-        retransmissions: int = 6,
-    ):
-        async def request():
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericDTTOpcode.GENERIC_DTT_SET_UNACKNOWLEDGED,
-                params=dict(
-                    transition_time=transition_time,
-                ),
-            )
-
-            return await ret
-
-        await self.repeat(
-            request,
-            retransmissions=retransmissions,
-            send_interval=send_interval,
-        )
-
-
-    async def get_dtt_status(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[float, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=GenericDTTOpcode.GENERIC_DTT_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = GenericDTTOpcode.GENERIC_DTT_STATUS
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-
-class GenericPowerOnOffServer(Model):
-    MODEL_ID = (None, 0x1006)
-    OPCODES = {
-        GenericPowerOnOffOpcode.GENERIC_ON_POWERUP_GET,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-
-class GenericPowerOnOffSetupServer(Model):
-    MODEL_ID = (None, 0x1007)
-    OPCODES = {
-        GenericPowerOnOffSetupOpcode.GENERIC_ON_POWERUP_SET,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-
-class GenericPowerOnOffClient(Model):
-    MODEL_ID = (None, 0x1008)
-    OPCODES = {
-        GenericPowerOnOffOpcode.GENERIC_ON_POWERUP_STATUS,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-    async def set_on_power_up(
-        self,
-        destination: int,
-        app_index: int,
-        on_power_up: GenericOnPowerUp,
-        send_interval: float = 0.07,
-        timeout: Optional[float] = None,
-    ) -> int:
-
-        status_opcode = GenericPowerOnOffOpcode.GENERIC_ON_POWERUP_STATUS
-
-        status = self.expect_app(
-            destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(on_power_up=on_power_up),
-        )
-
-        async def request():
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericPowerOnOffSetupOpcode.GENERIC_ON_POWERUP_SET,
-                params=dict(
-                    on_power_up=on_power_up,
-                ),
-            )
-
-            return await ret
-
-        status = await self.query(
-            request, status, send_interval=send_interval, timeout=timeout or 1
-        )
-
-        return status[status_opcode.name.lower()]["on_power_up"]
-
-
-    async def set_on_power_up_unack(
-        self,
-        destination: int,
-        app_index: int,
-        on_power_up: GenericOnPowerUp,
-        *,
-        send_interval: float = 0.1,
-        retransmissions: int = 6,
-    ):
-        async def request():
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericPowerOnOffSetupOpcode.GENERIC_ON_POWERUP_SET_UNACKNOWLEDGED,
-                params=dict(
-                    on_power_up=on_power_up,
-                ),
-            )
-
-            return await ret
-
-        await self.repeat(
-            request,
-            retransmissions=retransmissions,
-            send_interval=send_interval,
-        )
-
-
-    async def get_on_power_up(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=GenericPowerOnOffOpcode.GENERIC_ON_POWERUP_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = GenericPowerOnOffOpcode.GENERIC_ON_POWERUP_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=0,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-
 class SceneClient(Model):
     MODEL_ID = (None, 0x1205)
     OPCODES = {
@@ -1796,530 +1343,6 @@ class SceneClient(Model):
         }
 
         status_opcode = SceneOpcode.SCENE_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=0,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-
-class GenericLevelClient(Model):
-    MODEL_ID = (None, 0x1003)
-    OPCODES = {
-        GenericLevelOpcode.GENERIC_LEVEL_STATUS,
-        GenericLevelOpcode.GENERIC_LEVEL_GET,
-        GenericLevelOpcode.GENERIC_LEVEL_SET,
-        GenericLevelOpcode.GENERIC_LEVEL_SET_UNACKNOWLEDGED,
-        GenericLevelOpcode.GENERIC_DELTA_SET,
-        GenericLevelOpcode.GENERIC_DELTA_SET_UNACKNOWLEDGED,
-        GenericLevelOpcode.GENERIC_MOVE_SET,
-        GenericLevelOpcode.GENERIC_MOVE_SET_UNACKNOWLEDGED,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-    async def set_level(
-        self,
-        destination: int,
-        app_index: int,
-        level: int,
-        delay: float = 0.0,
-        transition_time: float = 0,
-        send_interval: float = 0.07,
-        timeout: Optional[float] = None,
-    ):
-        status_opcode = GenericLevelOpcode.GENERIC_LEVEL_STATUS
-        status = self.expect_app(
-            source=destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        tid = self.tid()
-        current_delay = delay
-
-        async def request():
-            nonlocal current_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericLevelOpcode.GENERIC_LEVEL_SET,
-                params=dict(
-                    level=level,
-                    tid=tid,
-                    transition_time=transition_time,
-                    delay=current_delay,
-                ),
-            )
-            current_delay = max(0.0, current_delay - send_interval)
-
-            return await ret
-
-        status = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 1
-        )
-
-        return status[status_opcode.name.lower()]
-
-
-    async def set_level_unack(
-        self,
-        destination: int,
-        app_index: int,
-        level: int,
-        delay: float = 0.0,
-        transition_time: float = 0,
-        send_interval: float = 0.07,
-        retransmissions: int = 6,
-    ):
-        tid = self.tid()
-        current_delay = delay
-
-        async def request():
-            nonlocal current_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericLevelOpcode.GENERIC_LEVEL_SET_UNACKNOWLEDGED,
-                params=dict(
-                    level=level,
-                    tid=tid,
-                    delay=current_delay,
-                    transition_time=transition_time,
-                ),
-            )
-            current_delay = max(0.0, current_delay - send_interval)
-
-            return await ret
-
-        await self.repeat(
-            request,
-            retransmissions=retransmissions,
-            send_interval=send_interval,
-        )
-
-
-    async def delta_set(
-        self,
-        destination: int,
-        app_index: int,
-        delta_level: int,
-        delay: float = 0.0,
-        transition_time: float = 0,
-        send_interval: float = 0.07,
-        timeout: Optional[float] = None,
-    ):
-        status_opcode = GenericLevelOpcode.GENERIC_LEVEL_STATUS
-        status = self.expect_app(
-            source=destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        tid = self.tid()
-        current_delay = delay
-
-        async def request():
-            nonlocal current_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericLevelOpcode.GENERIC_DELTA_SET,
-                params=dict(
-                    delta_level=delta_level,
-                    tid=tid,
-                    transition_time=transition_time,
-                    delay=current_delay,
-                ),
-            )
-            current_delay = max(0.0, current_delay - send_interval)
-
-            return await ret
-
-        status = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 1
-        )
-
-        return status[status_opcode.name.lower()]
-
-
-    async def move_set(
-        self,
-        destination: int,
-        app_index: int,
-        delta_level: int,
-        delay: float = 0.0,
-        transition_time: float = 0,
-        send_interval: float = 0.07,
-        timeout: Optional[float] = None,
-    ):
-        status_opcode = GenericLevelOpcode.GENERIC_LEVEL_STATUS
-        status = self.expect_app(
-            source=destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        tid = self.tid()
-        current_delay = delay
-
-        async def request():
-            nonlocal current_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=GenericLevelOpcode.GENERIC_MOVE_SET,
-                params=dict(
-                    delta_level=delta_level,
-                    tid=tid,
-                    transition_time=transition_time,
-                    delay=current_delay,
-                ),
-            )
-            current_delay = max(0.0, current_delay - send_interval)
-
-            return await ret
-
-        status = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 1
-        )
-
-        return status[status_opcode.name.lower()]
-
-
-    async def get_level_status(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=GenericLevelOpcode.GENERIC_LEVEL_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = GenericLevelOpcode.GENERIC_LEVEL_STATUS
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-
-class LightLightnessServer(Model):
-    MODEL_ID = (None, 0x1300)
-    OPCODES = {
-        LightLightnessOpcode.LIGHT_LIGHTNESS_GET,
-        LightLightnessOpcode.LIGHT_LIGHTNESS_SET,
-        LightLightnessOpcode.LIGHT_LIGHTNESS_SET_UNACKNOWLEDGED,
-        LightLightnessOpcode.LIGHT_LIGHTNESS_STATUS,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-
-class LightLightnessSetupServer(Model):
-    MODEL_ID = (None, 0x1301)
-    OPCODES = {
-        LightLightnessSetupOpcode.LIGHT_LIGHTNESS_SETUP_DEFAULT_SET,
-        LightLightnessSetupOpcode.LIGHT_LIGHTNESS_SETUP_DEFAULT_SET_UNACKNOWLEDGED,
-        LightLightnessSetupOpcode.LIGHT_LIGHTNESS_SETUP_RANGE_SET,
-        LightLightnessSetupOpcode.LIGHT_LIGHTNESS_SETUP_RANGE_SET_UNACKNOWLEDGED,
-    }
-    SUBSCRIBE = True
-
-
-class LightLightnessClient(Model):
-    MODEL_ID = (None, 0x1302)
-    OPCODES = {
-        LightLightnessOpcode.LIGHT_LIGHTNESS_STATUS,
-        LightLightnessOpcode.LIGHT_LIGHTNESS_RANGE_STATUS,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-    async def set_lightness_range_unack(
-        self,
-        destination: int,
-        app_index: int,
-        min_lightness: int,
-        max_lightness: int,
-        *,
-        retransmissions: int = 6,
-        send_interval: float = 0.075,
-    ) -> None:
-        async def request():
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=LightLightnessSetupOpcode.LIGHT_LIGHTNESS_SETUP_RANGE_SET_UNACKNOWLEDGED,
-                params=dict(
-                    range_min=min_lightness,
-                    range_max=max_lightness,
-                ),
-            )
-
-            return await ret
-
-        await self.repeat(
-            request, retransmissions=retransmissions, send_interval=send_interval
-        )
-
-    async def set_lightness_range(
-        self,
-        destination: int,
-        app_index: int,
-        min_lightness: int,
-        max_lightness: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Optional[Any]:
-        request = partial(
-            self.send_app,
-            destination=destination,
-            app_index=app_index,
-            opcode=LightLightnessSetupOpcode.LIGHT_LIGHTNESS_SETUP_RANGE_SET,
-            params=dict(
-                range_min=min_lightness, range_max=max_lightness, tid=self.tid()
-            ),
-        )
-
-        status_opcode = LightLightnessOpcode.LIGHT_LIGHTNESS_RANGE_STATUS
-
-        status = self.expect_app(
-            source=destination,
-            app_index=0,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        result = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 2.0,
-        )
-
-        return result[status_opcode.name.lower()]
-
-    async def get_lightness_range(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.075,
-        timeout: Optional[float] = None,
-    ) -> None:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightLightnessOpcode.LIGHT_LIGHTNESS_RANGE_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightLightnessOpcode.LIGHT_LIGHTNESS_RANGE_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-    async def get_lightness(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightLightnessOpcode.LIGHT_LIGHTNESS_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightLightnessOpcode.LIGHT_LIGHTNESS_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=0,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-    async def set_lightness_unack(
-        self,
-        destination: int,
-        app_index: int,
-        lightness: int,
-        transition_time: float = 0,
-        *,
-        delay: float = 0.5,
-        retransmissions: int = 6,
-        send_interval: float = 0.075,
-    ) -> None:
-        tid = self.tid()
-        remaining_delay = delay
-
-        async def request():
-            nonlocal remaining_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=LightLightnessOpcode.LIGHT_LIGHTNESS_SET_UNACKNOWLEDGED,
-                params=dict(
-                    lightness=lightness,
-                    delay=remaining_delay,
-                    tid=tid,
-                    transition_time=transition_time,
-                ),
-            )
-            remaining_delay = max(0.0, remaining_delay - send_interval)
-
-            return await ret
-
-        await self.repeat(
-            request, retransmissions=retransmissions, send_interval=send_interval
-        )
-
-    async def set_lightness(
-        self,
-        nodes: Sequence[int],
-        lightness: int,
-        app_index: int,
-        transition_time: float = 0,
-        *,
-        delay: float = 0.5,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        print("transition_time2=%f" % transition_time)
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightLightnessOpcode.LIGHT_LIGHTNESS_SET,
-                params=dict(
-                    lightness=lightness,
-                    tid=self.tid(),
-                    delay=0,
-                    transition_time=transition_time),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightLightnessOpcode.LIGHT_LIGHTNESS_STATUS
 
         statuses = {
             node: self.expect_app(
@@ -2448,657 +1471,6 @@ class SensorClient(Model):
             node: None if isinstance(result, Exception) else result[status.name.lower()]
             for node, result in results.items()
         }
-
-
-class LightCTLClient(Model):
-    MODEL_ID = (None, 0x1305)
-    OPCODES = {
-        LightCTLOpcode.LIGHT_CTL_GET,
-        LightCTLOpcode.LIGHT_CTL_SET,
-        LightCTLOpcode.LIGHT_CTL_SET_UNACKNOWLEDGED,
-        LightCTLOpcode.LIGHT_CTL_STATUS,
-        LightCTLOpcode.LIGHT_CTL_TEMPERATURE_GET,
-        LightCTLOpcode.LIGHT_CTL_TEMPERATURE_RANGE_GET,
-        LightCTLOpcode.LIGHT_CTL_TEMPERATURE_RANGE_STATUS,
-        LightCTLOpcode.LIGHT_CTL_TEMPERATURE_SET,
-        LightCTLOpcode.LIGHT_CTL_TEMPERATURE_SET_UNACKNOWLEDGED,
-        LightCTLOpcode.LIGHT_CTL_TEMPERATURE_STATUS,
-        LightCTLOpcode.LIGHT_CTL_TEMPERATURE_DEFAULT_GET,
-        LightCTLOpcode.LIGHT_CTL_TEMPERATURE_DEFAULT_STATUS,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-    async def get_ctl(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightCTLOpcode.LIGHT_CTL_TEMPERATURE_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightCTLOpcode.LIGHT_CTL_TEMPERATURE_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-    async def set_ctl(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        ctl_temperature: int,
-        transition_time: float = 0,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightCTLOpcode.LIGHT_CTL_TEMPERATURE_SET,
-                params=dict(
-                    ctl_temperature=ctl_temperature,
-                    ctl_delta_uv=0,
-                    tid=self.tid(),
-                    delay=0,
-                    transition_time=transition_time
-                ),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightCTLOpcode.LIGHT_CTL_TEMPERATURE_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-
-
-##########################
-
-class LightHSLServer(Model):
-    MODEL_ID = (None, 0x1307)
-    OPCODES = {
-        LightHSLOpcode.LIGHT_HSL_SET,
-        LightHSLOpcode.LIGHT_HSL_SET_UNACKNOWLEDGED,
-        LightHSLOpcode.LIGHT_HSL_STATUS,
-        LightHSLOpcode.LIGHT_HSL_TARGET_GET,
-        LightHSLOpcode.LIGHT_HSL_TARGET_STATUS,
-        LightHSLOpcode.LIGHT_HSL_DEFAULT_GET,
-        LightHSLOpcode.LIGHT_HSL_DEFAULT_STATUS,
-        LightHSLOpcode.LIGHT_HSL_RANGE_GET,
-        LightHSLOpcode.LIGHT_HSL_RANGE_STATUS,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-class LightHSLSetupServer(Model):
-    MODEL_ID = (None, 0x1308)
-    OPCODES = {
-        LightHSLSetupOpcode.LIGHT_HSL_SETUP_DEFAULT_SET,
-        LightHSLSetupOpcode.LIGHT_HSL_SETUP_DEFAULT_SET_UNACKNOWLEDGED,
-        LightHSLSetupOpcode.LIGHT_HSL_SETUP_RANGE_SET,
-        LightHSLSetupOpcode.LIGHT_HSL_SETUP_RANGE_SET_UNACKNOWLEDGED,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-class LightHSLHueServer(Model):
-    MODEL_ID = (None, 0x130A)
-    OPCODES = {
-        LightHSLOpcode.LIGHT_HSL_HUE_GET,
-        LightHSLOpcode.LIGHT_HSL_HUE_SET,
-        LightHSLOpcode.LIGHT_HSL_HUE_SET_UNACKNOWLEDGED,
-        LightHSLOpcode.LIGHT_HSL_HUE_STATUS,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-class LightHSLSaturationServer(Model):
-    MODEL_ID = (None, 0x130B)
-    OPCODES = {
-        LightHSLOpcode.LIGHT_HSL_SATURATION_GET,
-        LightHSLOpcode.LIGHT_HSL_SATURATION_SET,
-        LightHSLOpcode.LIGHT_HSL_SATURATION_SET_UNACKNOWLEDGED,
-        LightHSLOpcode.LIGHT_HSL_SATURATION_STATUS,
-    }
-    PUBLISH = True
-    SUBSCRIBE = True
-
-class LightHSLClient(Model):
-    MODEL_ID = (None, 0x1309)
-    OPCODES = {
-        LightHSLOpcode.LIGHT_HSL_GET,
-        LightHSLOpcode.LIGHT_HSL_SET,
-        LightHSLOpcode.LIGHT_HSL_SET_UNACKNOWLEDGED,
-        LightHSLOpcode.LIGHT_HSL_STATUS,
-        LightHSLOpcode.LIGHT_HSL_TARGET_GET,
-        LightHSLOpcode.LIGHT_HSL_TARGET_STATUS,
-        LightHSLOpcode.LIGHT_HSL_DEFAULT_GET,
-        LightHSLSetupOpcode.LIGHT_HSL_SETUP_DEFAULT_SET,
-        LightHSLSetupOpcode.LIGHT_HSL_SETUP_DEFAULT_SET_UNACKNOWLEDGED,
-        LightHSLOpcode.LIGHT_HSL_DEFAULT_STATUS,
-        LightHSLOpcode.LIGHT_HSL_RANGE_GET,
-        LightHSLOpcode.LIGHT_HSL_RANGE_STATUS,
-        LightHSLSetupOpcode.LIGHT_HSL_SETUP_RANGE_SET,
-        LightHSLSetupOpcode.LIGHT_HSL_SETUP_RANGE_SET_UNACKNOWLEDGED,
-        LightHSLOpcode.LIGHT_HSL_HUE_GET,
-        LightHSLOpcode.LIGHT_HSL_HUE_SET,
-        LightHSLOpcode.LIGHT_HSL_HUE_SET_UNACKNOWLEDGED,
-        LightHSLOpcode.LIGHT_HSL_HUE_STATUS,
-        LightHSLOpcode.LIGHT_HSL_SATURATION_GET,
-        LightHSLOpcode.LIGHT_HSL_SATURATION_SET,
-        LightHSLOpcode.LIGHT_HSL_SATURATION_SET_UNACKNOWLEDGED,
-        LightHSLOpcode.LIGHT_HSL_SATURATION_STATUS,
-    }
-
-    async def get_status(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightHSLOpcode.LIGHT_HSL_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightHSLOpcode.LIGHT_HSL_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-    async def get_target(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightHSLOpcode.LIGHT_HSL_TARGET_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightHSLOpcode.LIGHT_HSL_TARGET_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-    async def set(
-        self,
-        destination: int,
-        app_index: int,
-        lightness: int,
-        hue: int,
-        saturation: int,
-        *,
-        delay: float = 0,
-        transition_time: float = 0,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        current_delay = delay
-        tid = self.tid()
-
-        status_opcode = LightHSLOpcode.LIGHT_HSL_STATUS
-
-        status = self.expect_app(
-            destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        async def request():
-            nonlocal current_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=LightHSLOpcode.LIGHT_HSL_SET,
-                params=dict(
-                    hsl_lightness=lightness,
-                    hsl_hue=hue,
-                    hsl_saturation=saturation,
-                    tid=tid,
-                    transition_time=transition_time,
-                    delay=current_delay,
-                ),
-            )
-            current_delay = max(0.0, current_delay - send_interval)
-
-            return await ret
-
-        result = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 0.5,
-        )
-
-        return (None if isinstance(result, Exception)
-            else result[status_opcode.name.lower()])
-
-    async def get_hue(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightHSLOpcode.LIGHT_HSL_HUE_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightHSLOpcode.LIGHT_HSL_HUE_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-    async def set_hue(
-        self,
-        destination: int,
-        app_index: int,
-        hue: int,
-        *,
-        delay: float = 0,
-        transition_time: float = 0,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        current_delay = delay
-        tid = self.tid()
-
-        status_opcode = LightHSLOpcode.LIGHT_HSL_HUE_STATUS
-
-        status = self.expect_app(
-            destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        async def request():
-            nonlocal current_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=LightHSLOpcode.LIGHT_HSL_HUE_SET,
-                params=dict(
-                    hue=hue,
-                    tid=tid,
-                    transition_time=transition_time,
-                    delay=current_delay,
-                ),
-            )
-            current_delay = max(0.0, current_delay - send_interval)
-
-            return await ret
-
-        result = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 0.5,
-        )
-
-        return (None if isinstance(result, Exception)
-            else result[status_opcode.name.lower()])
-
-    async def get_saturation(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightHSLOpcode.LIGHT_HSL_SATURATION_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightHSLOpcode.LIGHT_HSL_SATURATION_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-    async def set_saturation(
-        self,
-        destination: int,
-        app_index: int,
-        saturation: int,
-        *,
-        delay: float = 0,
-        transition_time: float = 0,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        current_delay = delay
-        tid = self.tid()
-
-        status_opcode = LightHSLOpcode.LIGHT_HSL_SATURATION_STATUS
-
-        status = self.expect_app(
-            destination,
-            app_index=app_index,
-            destination=None,
-            opcode=status_opcode,
-            params=dict(),
-        )
-
-        async def request():
-            nonlocal current_delay
-            ret = self.send_app(
-                destination,
-                app_index=app_index,
-                opcode=LightHSLOpcode.LIGHT_HSL_SATURATION_SET,
-                params=dict(
-                    saturation=saturation,
-                    tid=tid,
-                    transition_time=transition_time,
-                    delay=current_delay,
-                ),
-            )
-            current_delay = max(0.0, current_delay - send_interval)
-
-            return await ret
-
-        result = await self.query(
-            request,
-            status,
-            send_interval=send_interval,
-            timeout=timeout or 0.5,
-        )
-
-        return (None if isinstance(result, Exception)
-            else result[status_opcode.name.lower()])
-
-    async def get_range(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightHSLOpcode.LIGHT_HSL_RANGE_GET,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightHSLOpcode.LIGHT_HSL_RANGE_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None if isinstance(result, Exception)
-                else result[status_opcode.name.lower()]
-                    for node, result in results.items()
-        }
-
-    async def set_range(
-        self,
-        nodes: Sequence[int],
-        app_index: int,
-        *,
-        hue_range_min: int,
-        hue_range_max: int,
-        saturation_range_min: int,
-        saturation_range_max: int,
-        send_interval: float = 0.1,
-        timeout: Optional[float] = None,
-    ) -> Dict[int, Optional[Any]]:
-        requests = {
-            node: partial(
-                self.send_app,
-                node,
-                app_index=app_index,
-                opcode=LightHSLSetupOpcode.LIGHT_HSL_SETUP_RANGE_SET,
-                params=dict(
-                    hue_range_min=hue_range_min,
-                    hue_range_max=hue_range_max,
-                    saturation_range_min=saturation_range_min,
-                    saturation_range_max=saturation_range_max,
-                    ctl_delta_uv=0,
-                    tid=self.tid()
-                ),
-            )
-            for node in nodes
-        }
-
-        status_opcode = LightHSLOpcode.LIGHT_HSL_RANGE_STATUS
-
-        statuses = {
-            node: self.expect_app(
-                node,
-                app_index=app_index,
-                destination=None,
-                opcode=status_opcode,
-                params=dict(),
-            )
-            for node in nodes
-        }
-
-        results = await self.bulk_query(
-            requests,
-            statuses,
-            send_interval=send_interval,
-            timeout=timeout or len(nodes) * 0.5,
-        )
-
-        return {
-            node: None
-            if isinstance(result, Exception)
-            else result[status_opcode.name.lower()]
-            for node, result in results.items()
-        }
-
-
-
-
-#############################
-
-
-
 
 class GatewayConfigServer(Model):
     MODEL_ID = (0x0136, 0x001D)
